refactor: createSSEClientTransport
